Pull out ns picker to it's own component so we can swap it out if necessary
